Make helper class to build function extensions
